backend/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
kombu==5.3.1
magic-filter==1.0.11
multidict==6.0.4
prompt-toolkit==3.0.39
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
